Allow tokenization of tuple types
With solidity abi v2 it has become possible to pass structs directly as
arguments to functions. A struct containing types T0 and T1 is encoded
as a tuple (T0, T1).

In the current state of rust-web3 it is not possible to pass tuples
because of the way the tokenization related traits are set up. We have
Tokenizable which is implemented for basic types like i32 and for arrays
of basic types but not for tuples of Tokenizable. Tokenize is
implemented on tuples of Tokenizable but not tuples of Tokenize so
nesting of tuples is not possible.
With this setup is not possible to make tuples tokenizable because it
would be ambiguous whether such a tuple represents a collection of
function arguments (or return types) which are encoded as Vec<Token> or
a single tuple which is encoded as Token::Tuple. The compiler notices
this too and prevents such an implementation.

This PR fixes the problem by changing the traits to resolve the
ambiguity.